user_backend/requirements.txt
pylint==3.0.2
pytest==7.4.3
python-decouple==3.8
Django==4.1
djangorestframework==3.14.0
markdown==3.5.1
django-filter==23.3
psycopg[binary,pool]==3.1.14
psycopg2-binary==2.9.9
